Takahashi-Matsuyama done, Kou to finish, Strategy implemented
Added the Strategy Design logic and class (UserInput class currently broken because of it), changed Dijkstra' logic to save the weights of the edges it adds to its path and added Prim's MST algorithm. Also (hopefully) finished the Takahashi-Matsuyama algorithm implementation (testTakahashi is performing as expected), and started the Kou-Markowsky-Berman algorithm implementation. To-do: steps 4 and 5 of Kou's algorithm, fix UserInput according to Strategy and obviously add the third ST approximation method.